+++ example/pages/ptr/index.js
@@ -8,6 +8,8 @@ import {
     CellFooter
 } from '../../../src/index';
 import Page from '../../component/page';
+import momentLoader from './momentloader.svg';
+import './ptr.less';
 
 class PtrDemo extends React.Component {
 
@@ -23,7 +25,7 @@ class PtrDemo extends React.Component {
         return (
         <Page className="ptr" title="Pull To Refresh" subTitle="下拉刷新" >
             <div style={{
-                height: '500px',
+                height: '200px',
                 background: '#fff'
             }}>
                 <PullToRefresh
@@ -39,6 +41,64 @@ class PtrDemo extends React.Component {
                     }
                 >
                     <CellsTitle>List with link</CellsTitle>
+                    <Cells>
+                    {
+                        this.state.items.map( (item, i) => {
+                            return (
+                                <Cell href="javascript:;" key={i} access>
+                                    <CellBody>
+                                        {item}
+                                    </CellBody>
+                                    <CellFooter/>
+                                </Cell>
+                            )
+                        })
+                    }
+                    </Cells>
+
+                </PullToRefresh>
+            </div>
+            <br/>
+
+            <div style={{
+                height: '200px',
+                background: '#fff'
+            }}>
+                <PullToRefresh
+                    loaderDefaultIcon={
+                        (progress) => {
+                            let style = {
+                                transform: `rotate(-${progress * 5}deg)`
+                            }
+                            return (
+                                <div style={{ flex: 1, padding: '5px' }}>
+                                    <img src={momentLoader} width="40px" style={style}/>
+                                </div>
+                            )
+                        }
+                    }
+                    loaderLoadingIcon={
+                        <div style={{
+                            flex: 1,
+                            padding: '5px',
+                        }}>
+                            <img src={momentLoader} width="40px" style={{
+                                animation: '0.4s spin infinite linear'
+                            }}/>
+                        </div>
+                    }
+                    onRefresh={
+                        resolve => {
+                            //mock add item after 1s and then resolve
+                            setTimeout(()=>{
+                                this.setState({
+                                    items: this.state.items.concat([this.state.items.length + 1])
+                                }, ()=> resolve())
+                            }, 1000)
+                        }
+                    }
+                >
+                    <CellsTitle>Moment Loader Example</CellsTitle>
                     <Cells>
                     {
                         this.state.items.map( (item, i) => {